refactor: rename keeltest folder
