+++ go.mod
@@ -38,6 +38,7 @@ require (
 	go.temporal.io/api v1.8.0
 	go.temporal.io/sdk v1.15.0
 	go.temporal.io/sdk/contrib/opentelemetry v0.1.0
+	go.uber.org/goleak v1.3.0
 	go.uber.org/zap v1.27.0
 	golang.org/x/crypto v0.22.0
 	golang.org/x/sync v0.7.0
