+++ keeltest/assert/assertions.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/foomo/keel/keeltest"
 	"github.com/foomo/keel/log"
-	keeltestutil "github.com/foomo/keel/test/util"
 	"github.com/stretchr/testify/assert"
 	"github.com/tidwall/pretty"
 )
@@ -26,7 +26,7 @@ func New(t *testing.T) *Assertions { //nolint:thelper
 
 func (a *Assertions) InlineEqual(actual interface{}, msgAndArgs ...interface{}) bool {
 	a.t.Helper()
-	expected, ok := keeltestutil.Inline(a.t, 2, "%v", actual)
+	expected, ok := keeltest.Inline(a.t, 2, "%v", actual)
 	if ok {
 		return assert.Equal(a.t, expected, fmt.Sprintf("%v", actual), msgAndArgs...)
 	} else {
@@ -42,7 +42,7 @@ func (a *Assertions) InlineJSONEq(actual interface{}, msgAndArgs ...interface{})
 		a.t.Fatal("failed to marshal json", log.FError(err))
 	}
 
-	expected, ok := keeltestutil.Inline(a.t, 2, string(actualBytes))
+	expected, ok := keeltest.Inline(a.t, 2, string(actualBytes))
 	if ok {
 		return assert.Equal(a.t, string(pretty.Pretty([]byte(expected))), string(pretty.Pretty(actualBytes)), msgAndArgs...)
 	} else {
