+++ keeltest/client.go
