+++ keeltest/require/assertions.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/foomo/keel/keeltest"
 	"github.com/foomo/keel/log"
-	keeltestutil "github.com/foomo/keel/test/util"
 	"github.com/stretchr/testify/require"
 	"github.com/tidwall/pretty"
 )
@@ -26,7 +26,7 @@ func New(t *testing.T) *Assertions { //nolint:thelper
 
 func (a *Assertions) InlineEqual(actual interface{}, msgAndArgs ...interface{}) {
 	a.t.Helper()
-	if expected, ok := keeltestutil.Inline(a.t, 2, "%v", actual); ok {
+	if expected, ok := keeltest.Inline(a.t, 2, "%v", actual); ok {
 		require.Equal(a.t, expected, fmt.Sprintf("%v", actual), msgAndArgs...)
 	}
 }
@@ -38,7 +38,7 @@ func (a *Assertions) InlineJSONEq(actual interface{}, msgAndArgs ...interface{})
 	if err != nil {
 		a.t.Fatal("failed to marshal json", log.FError(err))
 	}
-	if expected, ok := keeltestutil.Inline(a.t, 2, string(actualBytes)); ok {
+	if expected, ok := keeltest.Inline(a.t, 2, string(actualBytes)); ok {
 		require.Equal(a.t, string(pretty.Pretty([]byte(expected))), string(pretty.Pretty(actualBytes)), msgAndArgs...)
 	}
 }
