+++ keeltest/require/require.go
@@ -5,15 +5,15 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/foomo/keel/keeltest"
 	"github.com/foomo/keel/log"
-	keeltestutil "github.com/foomo/keel/test/util"
 	"github.com/stretchr/testify/require"
 	"github.com/tidwall/pretty"
 )
 
 func InlineEqual(t *testing.T, actual interface{}, msgAndArgs ...interface{}) {
 	t.Helper()
-	if expected, ok := keeltestutil.Inline(t, 2, "%v", actual); ok {
+	if expected, ok := keeltest.Inline(t, 2, "%v", actual); ok {
 		require.Equal(t, expected, fmt.Sprintf("%v", actual), msgAndArgs...)
 	}
 }
@@ -25,7 +25,7 @@ func InlineJSONEq(t *testing.T, actual interface{}, msgAndArgs ...interface{}) {
 	if err != nil {
 		t.Fatal("failed to marshal json", log.FError(err))
 	}
-	if expected, ok := keeltestutil.Inline(t, 2, string(actualBytes)); ok {
+	if expected, ok := keeltest.Inline(t, 2, string(actualBytes)); ok {
 		require.Equal(t, string(pretty.Pretty([]byte(expected))), string(pretty.Pretty(actualBytes)), msgAndArgs...)
 	}
 }
