+++ keeltest/service.go
